Resolve visit URI against URI of last request
This is what a browser would do, and it allows users to set a
non-localhost hostname at the beginning of tests and get correct
cookie behavior. Previously, a cookie set in a request to
http://example.com/foo would not be available when kerodon followed a
relative link on that page because the absence of an explicit hostname
in the link would result in `localhost` being used as the hostname for
that request.