Enable true validation of :actual values in kerodon.unit.test
When test cases did not result in :error, the :actual values were not being
validated due to a misplaced parenthesis! Can you spot the error? Ha ha!

As part of this changeset, also fixed minor issues in expected :actual values.
I suppose no one caught these as they were not actually being executed?

Lastly, refactored check-report's map parameter to expect :valiator-fn key to
provide a function to validate the :actual value for :error cases. Overloading
the :actual key for both a function (in :error cases) and a value (otherwise)
was a bit confusing.